+++ HDF5/H5Fpublic.cs
@@ -23,7 +23,7 @@ using hid_t = System.Int32;
 using hsize_t = System.UInt64;
 using hssize_t = System.Int64;
 using htri_t = System.Int32;
-using size_t = System.UInt64;
+using size_t = System.IntPtr;
 using ssize_t = System.Int32;
 
 namespace HDF.PInvoke
